+++ CHANGES.rst
@@ -11,6 +11,7 @@ Changelog
   (TBD). Use IPWhois.lookup_whois() instead.
 - Added rate_limit_timeout parameter to Net.get_http_json(),
   IPWhois.lookup_rdap(), and RDAP.lookup(). New exception HTTPRateLimitError.
+- Fixed socket exception handling in Net.get_http_json() for Python 2.6.
 
 0.11.2 (2016-02-25)
 -------------------